push alive_api to a static lib
