+++ Source/Tools/alive_api_test.cpp
@@ -1,0 +1,8 @@
+#include "alive_api.hpp"
+#include "SDL.h"
+
+int main(int argc, char* argv[])
+{
+    auto ret = AliveAPI::ExportPathBinaryToJson("Output.json", "C:\\GOG Games\\Abes Exoddus\\MI.LVL", 1);
+    return ret.errorCode;
+}
